door-api/requirements.txt
Flask==1.1.1
schedule==0.6.0
RPi.GPIO==0.7.0
